accordian dropdowns

+++ porefection/src/components/Routine/index.js
@@ -5,11 +5,31 @@ function Routine({ routine, getRoutine }) {
     // remove product
     const removeProduct = (product) => {
 
-        const updatedRoutine = routine.filter((_, index) => index !== product);
+        let removeConfirmed = window.confirm('Are you sure you want to remove this product from your routine?');
 
-        localStorage.setItem('Porefection Skincare Routine', JSON.stringify(updatedRoutine));
+        if (removeConfirmed) {
 
-        getRoutine();
+            const updatedRoutine = routine.filter((_, index) => index !== product);
+            localStorage.setItem('Porefection Skincare Routine', JSON.stringify(updatedRoutine));
+            getRoutine();
+
+        };
+
+    };
+
+    const handleAccordian = (e) => {
+
+        let carrot = e.target.lastChild;
+
+        carrot.style.transform === 'scaleY(-1) translateY(-3px) rotate(45deg)'
+            ? carrot.style.transform = 'scaleY(1) translateY(-3px) rotate(45deg)'
+            : carrot.style.transform = 'scaleY(-1) translateY(-3px) rotate(45deg)';
+
+        let panel = e.target.nextElementSibling;
+
+        panel.style.maxHeight
+            ? panel.style.maxHeight = null
+            : panel.style.maxHeight = panel.scrollHeight + "px";
 
     };
 
@@ -31,11 +51,24 @@ function Routine({ routine, getRoutine }) {
 
                             <div className="product-details">
 
-                                <h2>{product.displayName}</h2>
-                                <h4>{product.brand.displayName}</h4>
+                                <div className="product-banner">
 
-                                <h3 className="product-detail-header">Suggested Use</h3>
-                                <ol>
+                                    <div className="product-headline">
+
+                                        <h2>{product.displayName}</h2>
+                                        <h4>{product.brand.displayName}</h4>
+
+                                    </div>
+
+                                    <div className="remove-product-btn" onClick={() => removeProduct(index)}>
+                                        <div className="remove-product-left"></div>
+                                        <div className="remove-product-right"></div>
+                                    </div>
+
+                                </div>
+
+                                <h3 className="product-detail-header" onClick={handleAccordian}>Suggested Use<span className="carrot" /></h3>
+                                <ol className="accordian">
                                     {product.suggestedUsage.replace(/<br>|<b>|<\/b>|/g, '').split('-').map((line, index) => {
                                         return (
                                             index > 0 && <li key={index} className="use-line">{line}</li>
@@ -43,25 +76,25 @@ function Routine({ routine, getRoutine }) {
                                     })}
                                 </ol>
 
-                                <h3 className="product-detail-header">Ingredient Breakdown</h3>
-                                {product.currentSku.ingredientDesc.replace(/, Inactive Ingredients/g, '<br>Inactive Ingredients').replace(/<br><\/b>/g, ':').replace(/<b>|<\/b>| &ndash/g, '').split('<br>').map((line, index) => {
-                                    while (line.charAt(0) === '-') {
-                                        line = line.substring(1);
-                                    }
-                                    return (
-                                        <div className="ingredient">
-                                            {line.split(':').map((item, i) => {
-                                                return (
-                                                    <p key={index + '.' + i} className="ingredient-line">{item}</p>
-                                                );
-                                            })}
-                                        </div>
-                                    );
-                                })}
+                                <h3 className="product-detail-header" onClick={handleAccordian}>Ingredient Breakdown<span className="carrot" /></h3>
+                                <ul className="accordian">
+                                    {product.currentSku.ingredientDesc.replace(/, Inactive Ingredients/g, '<br>Inactive Ingredients').replace(/<br><\/b>/g, ':').replace(/<b>|<\/b>| &ndash/g, '').split('<br>').map((line, index) => {
+                                        while (line.charAt(0) === '-') {
+                                            line = line.substring(1);
+                                        }
+                                        return (
+                                            <div className="ingredient">
+                                                {line.split(':').map((item, i) => {
+                                                    return (
+                                                        <p key={index + '.' + i} className="ingredient-line">{item}</p>
+                                                    );
+                                                })}
+                                            </div>
+                                        );
+                                    })}
+                                </ul>
 
                             </div>
-
-                            <div className="remove-product-btn" onClick={() => removeProduct(index)}>x</div>
 
                         </li>
